refactor: move missing-header-error to utils layer
