+++ src/utils/errors/index.js
@@ -1,7 +1,9 @@
 const MissingParamError = require('./missing-param-error')
 const InvalidParamError = require('./invalid-param-error')
+const MissingHeaderError = require('./missing-header-error')
 
 module.exports = {
   MissingParamError,
-  InvalidParamError
+  InvalidParamError,
+  MissingHeaderError
 }
